Added solution for Q200 - Number of Islands
There are multiple ways to skin the car - I chose the recursive solution as it was easiest to understand.

For 100k by 100k grid, I am still running out of memory, so there's a lot a scope for improvement there. I could just increase the heap size, but where's the fun in that!