Added Q 347 - Top K Frequent Elements in Array
Given an integer array `nums` and an integer `k`, return the `k` most frequent elements in any order.

Using Hashmap to first calculate the frequency. There are then multiple ways to sort this map - TreeMap, Priority Queue, LinkedHashMap with a custom comparator and calling Collections.sort() on it, etc.

I chose the Priority Queue method since this is the easiest.