Added Q658 - Find K closest elements in an array
Given a sorted integer array `arr`, two integers `k` and `x`, return the `k` closest integers to `x` in the array. The result should also be sorted in ascending order.

I initially used Priority Queue. I checked and there are way better ways to do it, since we were not utilizing the sorted aspect of the code.

Added a second code where we are indeed doing that using a mix of Binary Search and Sliding Window.