Added solution for Q743 - Network Delay Time
Using Dijktra's algorithm, find the shortest route to each node form starting point k.

Once you have the data distance[] which is 1 based, iterate from index 1 to the end of the array and find the max value in the array.
If max value is Integer.MAX_VALUE, we can safely assume that the node is unreachable.

return either max value or -1;